Add UTF-8 support for diplay() and movement; insert() not done yet.
A simple file with fire glyph (U+1f525) and LF.  NetBSD Curses
appears to have an issue displaying % position (phantom backspace)
and the cursor position on the LF following the wide character.
However, NCurses works as expected.

When the wide character is on the second line following an line of
ASCII the display and cursor motion across the LF works as expected.

Test cases needed for these two cases.